backend/requirements.txt
typer>=0.9
fastapi>=0.115
uvicorn>=0.23
pydantic>=2.7
pdfplumber>=0.10
Pillow>=10.0
pytesseract>=0.3.10
rich>=13.0
